refactor: clean up after server driven navigation feature
