+++ src/Apps/Components/__tests__/AppShell.jest.tsx
@@ -17,6 +17,13 @@ jest.mock("react-tracking", () => ({
   }),
 }))
 
+jest.mock("react-relay", () => ({
+  ...jest.requireActual("react-relay"),
+  ReactRelayContext: {
+    Provider: ({ children }) => children,
+  },
+}))
+
 jest.mock("Utils/Hooks/useAuthValidation")
 
 jest.mock("Components/Footer/FooterDownloadAppBanner", () => ({
@@ -24,6 +31,24 @@ jest.mock("Components/Footer/FooterDownloadAppBanner", () => ({
 }))
 
 describe("AppShell", () => {
+  const fetchMock = jest.fn().mockResolvedValue({
+    json: async () => ({
+      data: {
+        whatsNewNavigation: null,
+        artistsNavigation: null,
+        artworksNavigation: null,
+      },
+    }),
+  })
+
+  beforeAll(() => {
+    ;(global as any).fetch = fetchMock
+  })
+
+  beforeEach(() => {
+    fetchMock.mockClear()
+  })
+
   it("renders a Footer", async () => {
     const { ClientRouter } = await setupClientRouter({
       history: {
