+++ src/Components/NavBar/Menus/index.tsx
@@ -1,2 +1,1 @@
-export * from "./NavBarSubMenu"
 export * from "./NavBarUserMenu"
