+++ src/Components/NavBar/NavBarMobileMenu/NavBarMobileSubMenu.tsx
@@ -1,37 +1,72 @@
-import ChevronLeftIcon from "@artsy/icons/ChevronLeftIcon"
 import ChevronRightIcon from "@artsy/icons/ChevronRightIcon"
 import { Box, Flex, Separator, Text } from "@artsy/palette"
-import {
-  type LinkData,
-  type MenuData,
-  isMenuLinkData,
-} from "Components/NavBar/menuData"
 import type * as React from "react"
 import { useEffect, useRef } from "react"
+import { graphql, useFragment } from "react-relay"
 import {
   NavBarMobileMenuItemButton,
   NavBarMobileMenuItemLink,
 } from "./NavBarMobileMenuItem"
+import { NavBarMobileSubMenuBackButton } from "./NavBarMobileSubMenuBackButton"
 import { useNavBarMobileMenuNavigation } from "./NavBarMobileMenuNavigation"
 import { NavBarMobileMenuTransition } from "./NavBarMobileMenuTransition"
 import { useTrackingContextModule } from "./useTrackingContextModule"
 import { useNavBarTracking } from "../useNavBarTracking"
+import type { NavBarMobileSubMenu_navigationVersion$key } from "__generated__/NavBarMobileSubMenu_navigationVersion.graphql"
+
+const LETTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ".split("")
 
 interface NavBarMobileSubMenuProps {
-  menu: MenuData
+  navigationVersion: NavBarMobileSubMenu_navigationVersion$key
+  menuType: "whatsNew" | "artists" | "artworks"
   level?: number
 }
 
 export const NavBarMobileSubMenu: React.FC<
   React.PropsWithChildren<NavBarMobileSubMenuProps>
-> = ({ children, menu, level = 0 }) => {
+> = ({
+  children,
+  navigationVersion: navigationVersionProp,
+  menuType,
+  level = 0,
+}) => {
+  const navigationVersion = useFragment(
+    NAVIGATION_VERSION_FRAGMENT,
+    navigationVersionProp,
+  )
+
+  if (!navigationVersion?.items) {
+    return null
+  }
+
+  return (
+    <NavBarMobileSubMenuDrilldown
+      title={children?.toString() || ""}
+      navigationData={navigationVersion}
+      menuType={menuType}
+      level={level}
+    />
+  )
+}
+
+interface NavBarMobileSubMenuDrilldownProps {
+  title: string
+  navigationData: any // Plain unwrapped data
+  menuType: "whatsNew" | "artists" | "artworks"
+  level: number
+}
+
+const NavBarMobileSubMenuDrilldown: React.FC<
+  NavBarMobileSubMenuDrilldownProps
+> = ({ title, navigationData, menuType, level }) => {
   const tracking = useNavBarTracking()
   const { path, push } = useNavBarMobileMenuNavigation()
   const contextModule = useTrackingContextModule()
 
   const hasTrackedRef = useRef(false)
   const timeoutRef = useRef<NodeJS.Timeout | null>(null)
-  const isOpen = !!path?.includes(menu.title)
+  // Check if this menu is open by matching title at the correct level
+  const isOpen = path?.[level] === title
 
   // Track when drilldown menu becomes visible (once per page load, with 500ms delay)
   // biome-ignore lint/correctness/useExhaustiveDependencies: only track once per page load based on open state
@@ -40,7 +75,7 @@ export const NavBarMobileSubMenu: React.FC<
       // Only fire if menu stays open for at least 500ms
       timeoutRef.current = setTimeout(() => {
         tracking.navigationDropdownViewed({
-          navigationItem: menu.title,
+          navigationItem: title,
           level: level,
           interactionType: "drilldown",
         })
@@ -63,16 +98,16 @@ export const NavBarMobileSubMenu: React.FC<
     <>
       <NavBarMobileMenuItemButton
         onClick={() => {
-          push(menu.title)
+          push(title)
 
           tracking.clickedMobileSubMenuOpen({
             contextModule,
-            subject: menu.title,
+            subject: title,
           })
         }}
         aria-expanded={false}
       >
-        {children}
+        {title}
 
         <ChevronRightIcon
           fill="mono60"
@@ -85,22 +120,20 @@ export const NavBarMobileSubMenu: React.FC<
 
       <NavBarMobileSubMenuPanel
         isOpen={isOpen}
-        title={menu.title}
-        links={menu.links}
-        parentNavigationItem={menu.title}
+        title={title}
+        navigationVersion={navigationData}
+        menuType={menuType}
         level={level}
       />
     </>
   )
 }
 
-const LETTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ".split("")
-
 interface NavBarMobileSubMenuPanelProps {
   isOpen: boolean
   title: string
-  links: LinkData[]
-  parentNavigationItem: string
+  navigationVersion: any // Already fragmented
+  menuType: "whatsNew" | "artists" | "artworks"
   level: number
   showBacknav?: boolean
 }
@@ -110,17 +143,33 @@ const NavBarMobileSubMenuPanel: React.FC<
 > = ({
   isOpen,
   title,
-  links,
-  parentNavigationItem,
+  navigationVersion,
+  menuType,
   level,
   showBacknav = true,
 }) => {
-  const isArtistsMenu = title === "Artists"
+  const tracking = useNavBarTracking()
+  const contextModule = useTrackingContextModule()
+
+  const handleClick = (
+    _event: React.MouseEvent<HTMLAnchorElement, MouseEvent>,
+    itemTitle: string,
+    href: string,
+    sectionTitle: string,
+  ) => {
+    tracking.clickedNavigationDropdownItem({
+      contextModule,
+      parentNavigationItem: title,
+      dropdownGroup: sectionTitle,
+      subject: itemTitle,
+      destinationPath: href,
+    })
+  }
 
   return (
     <NavBarMobileMenuTransition isOpen={isOpen}>
       <Flex position="relative" alignItems="stretch" height={60}>
-        {showBacknav && <NavBarMobileSubMenuBack />}
+        {showBacknav && <NavBarMobileSubMenuBackButton />}
 
         <Text variant="sm" color="mono100" display="flex" m="auto">
           {title}
@@ -130,103 +179,113 @@ const NavBarMobileSubMenuPanel: React.FC<
         {showBacknav && <Box size={60} />}
       </Flex>
 
-      {links.map((link, i) => {
+      {navigationVersion.items.map((section, sectionIndex) => {
+        if (!section?.title) return null
+
         return (
-          <NavBarMobileSubMenuItem
-            key={i}
-            link={link}
-            parentNavigationItem={parentNavigationItem}
-            level={level}
-          />
+          <Box key={sectionIndex}>
+            {/* Render each section as a drill-down button */}
+            <NavBarMobileSubMenuItem
+              item={section}
+              parentNavigationItem={title}
+              sectionTitle=""
+              level={level + 1}
+              menuType={menuType}
+            />
+          </Box>
         )
       })}
 
-      {isArtistsMenu && (
+      {/* View All links */}
+      {menuType === "artists" && level === 0 && (
         <>
           <Separator my={1} />
-
-          <Flex flexWrap="wrap" justifyContent="center">
-            {LETTERS.map(letter => {
-              return (
-                <NavBarMobileMenuItemLink
-                  key={letter}
-                  width={60}
-                  px={0}
-                  justifyContent="center"
-                  to={`/artists/artists-starting-with-${letter.toLowerCase()}`}
-                  title={`View artists starting with “${letter}”`}
-                >
-                  {letter}
-                </NavBarMobileMenuItemLink>
-              )
-            })}
-          </Flex>
+          <NavBarMobileMenuItemLink
+            to="/artists"
+            onClick={e => handleClick(e, "View All Artists", "/artists", "")}
+          >
+            View All Artists
+          </NavBarMobileMenuItemLink>
+          <Separator my={1} />
         </>
+      )}
+
+      {menuType === "artworks" && level === 0 && (
+        <>
+          <Separator my={1} />
+          <NavBarMobileMenuItemLink
+            to="/collect"
+            onClick={e => handleClick(e, "View All Artworks", "/collect", "")}
+          >
+            View All Artworks
+          </NavBarMobileMenuItemLink>
+          <Separator my={1} />
+        </>
+      )}
+
+      {/* A-Z letters for Artists menu */}
+      {menuType === "artists" && level === 0 && (
+        <Flex flexWrap="wrap" justifyContent="center">
+          {LETTERS.map(letter => {
+            return (
+              <NavBarMobileMenuItemLink
+                key={letter}
+                width={60}
+                px={0}
+                justifyContent="center"
+                to={`/artists/artists-starting-with-${letter.toLowerCase()}`}
+                title={`View artists starting with "${letter}"`}
+                onClick={e =>
+                  handleClick(
+                    e,
+                    letter,
+                    `/artists/artists-starting-with-${letter.toLowerCase()}`,
+                    "Browse by name",
+                  )
+                }
+              >
+                {letter}
+              </NavBarMobileMenuItemLink>
+            )
+          })}
+        </Flex>
       )}
     </NavBarMobileMenuTransition>
   )
 }
 
-export const NavBarMobileSubMenuBack: React.FC<
-  React.PropsWithChildren<unknown>
-> = () => {
-  const tracking = useNavBarTracking()
-  const { pop } = useNavBarMobileMenuNavigation()
-  const contextModule = useTrackingContextModule()
-
-  return (
-    <NavBarMobileMenuItemButton
-      width={60}
-      height={60}
-      display="flex"
-      py={0}
-      px={0}
-      aria-label="Back"
-      onClick={() => {
-        tracking.clickedMobileSubMenuBack({ contextModule })
-
-        pop()
-      }}
-    >
-      <ChevronLeftIcon
-        fill="mono100"
-        height={14}
-        width={14}
-        m="auto"
-        aria-hidden
-      />
-    </NavBarMobileMenuItemButton>
-  )
-}
-
 interface NavBarMobileSubMenuItemProps {
-  link: LinkData
+  item: any // NavigationItem from server
   parentNavigationItem: string
+  sectionTitle: string
   level: number
+  menuType: "whatsNew" | "artists" | "artworks"
 }
 
-export const NavBarMobileSubMenuItem: React.FC<
-  React.PropsWithChildren<NavBarMobileSubMenuItemProps>
-> = ({ link, parentNavigationItem, level }) => {
+const NavBarMobileSubMenuItem: React.FC<NavBarMobileSubMenuItemProps> = ({
+  item,
+  parentNavigationItem,
+  sectionTitle,
+  level,
+  menuType,
+}) => {
   const tracking = useNavBarTracking()
   const contextModule = useTrackingContextModule()
 
-  if (isMenuLinkData(link)) {
+  // If item has children, it's a nested submenu - use the drilldown component
+  if (item.children && item.children.length > 0) {
     return (
-      <>
-        <NavBarMobileSubMenu menu={link.menu} level={level + 1}>
-          {link.text}
-        </NavBarMobileSubMenu>
-
-        {link.dividerBelow && <Separator my={1} />}
-      </>
+      <NavBarMobileSubMenuDrilldown
+        title={item.title}
+        navigationData={{ items: item.children }}
+        menuType={menuType}
+        level={level}
+      />
     )
   }
 
-  // Skip visual components (like articles) in mobile menu
-  if ("type" in link) {
-    return null
-  }
+  // Otherwise it's a regular link
+  if (!item.href) return null
 
   const handleClick = (
     event: React.MouseEvent<HTMLAnchorElement, MouseEvent>,
@@ -234,20 +293,29 @@ export const NavBarMobileSubMenuItem: React.FC<
     tracking.clickedNavigationDropdownItem({
       contextModule,
       parentNavigationItem,
-      subject: link.text,
-      destinationPath: link.href,
+      dropdownGroup: sectionTitle,
+      subject: item.title,
+      destinationPath: item.href,
     })
-
-    link.onClick?.(event)
   }
 
   return (
-    <>
-      <NavBarMobileMenuItemLink to={link.href} onClick={handleClick}>
-        {link.text}
-      </NavBarMobileMenuItemLink>
-
-      {link.dividerBelow && <Separator my={1} />}
-    </>
+    <NavBarMobileMenuItemLink to={item.href} onClick={handleClick}>
+      {item.title}
+    </NavBarMobileMenuItemLink>
   )
 }
+
+const NAVIGATION_VERSION_FRAGMENT = graphql`
+  fragment NavBarMobileSubMenu_navigationVersion on NavigationVersion {
+    items {
+      title
+      position
+      children {
+        title
+        href
+        position
+      }
+    }
+  }
+`
