+++ src/Server/bootstrapSharify.ts
@@ -45,7 +45,6 @@ export const bootstrapSharify = () => {
       "ENABLE_NEW_AUCTIONS_FILTER",
       "ENABLE_PREFETCH",
       "ENABLE_QUERY_BATCHING",
-      "ENABLE_SERVER_DRIVEN_NAVIGATION",
       "ENABLE_SSR_STREAMING",
       "ENABLE_WEB_CRAWLING",
       "ENABLE_WEB_VITALS_LOGGING",
