+++ src/Server/config.ts
@@ -52,7 +52,6 @@ export const ENABLE_CONVERSATIONS_MESSAGES_AUTO_REFRESH: any = true
 export const ENABLE_NEW_AUCTIONS_FILTER: any = false
 export const ENABLE_PREFETCH: any = true
 export const ENABLE_QUERY_BATCHING: any = false
-export const ENABLE_SERVER_DRIVEN_NAVIGATION: any = false
 export const ENABLE_SSR_STREAMING: any = false
 export const ENABLE_WEB_CRAWLING: any = false
 export const ENABLE_WEB_VITALS_LOGGING: any = false
